requirements.txt
# Core dependencies
PySide6>=6.0.0
tensorflow>=2.7.0
Pillow>=9.0.0
numpy>=1.19.2
mysql-connector-python>=8.0.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-qt>=4.0.0

# Development tools
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950